requirements.txt
psycopg2-binary>=2.9.0
google-generativeai>=0.3.0
numpy>=1.24.0
pgvector>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
ollama>=0.1.0
streamlit>=1.28.0